requirements.txt
pydantic
fastapi
sqlalchemy
uvicorn[standard]
psycopg2-binary
alembic
python-dotenv